Interview Final/requirements.txt
langchain-community>=0.0.10
langchain-openai>=0.0.3
pydantic
aiohttp>=3.9.0
python-dotenv
streamlit